+++ src/TicketModalController/TicketModalController.js
@@ -43,6 +43,15 @@ class TicketModalController extends Component {
     fetchAvailableTicketsCancel();
   }
 
+  getAvailablePoolTicketsQty = () => {
+    const { poolId } = this.state;
+    const { availableTickets: { ticketPools } } = this.props;
+    const ticketPool = this.getTicketPoolById(ticketPools, poolId) || {};
+    const { availableTicketsNumber } = ticketPool;
+
+    return Number.isInteger(availableTicketsNumber) ? availableTicketsNumber : null;
+  };
+
   getInitialActiveStep = (cartItem) => {
     const { sightEvent } = this.props;
     const { ticketPoolDefinitions } = sightEvent || {};
@@ -106,6 +115,19 @@ class TicketModalController extends Component {
     return null;
   };
 
+  getTicketAvailabilityText = (totalTicketsNumber, selectedTicketsNumber, options = {}) => {
+    const t = options.t || (() => '')();
+    const availableTicketsNumber = totalTicketsNumber - selectedTicketsNumber;
+    let availability = t('ticketModal:ticketAvailability:unlimited');
+
+    if (totalTicketsNumber === 0) {
+      availability = t('ticketModal:ticketAvailability:unavailable');
+    } else if (totalTicketsNumber > 0) {
+      availability = t('ticketModal:ticketAvailability:available', { availableTicketsNumber });
+    }
+
+    return availability;
+  };
 
   getTicketDefinitions = (poolId) => {
     const { availableTickets: { ticketPools } } = this.props;
@@ -445,6 +467,7 @@ class TicketModalController extends Component {
     return children({
       ...props,
       ...this.state,
+      availablePoolTicketsQty: this.getAvailablePoolTicketsQty(),
       totalPrice: this.getTotalPrice(),
       totalTicketsQty: this.getTotalTicketsQty(),
       handleAgreementChange: this.handleAgreementChange,
@@ -453,6 +476,7 @@ class TicketModalController extends Component {
       handlePoolChange: this.handlePoolChange,
       handlePrevButtonClick: this.handlePrevButtonClick,
       handleNextButtonClick: this.handleNextButtonClick,
+      getTicketAvailabilityText: this.getTicketAvailabilityText,
       getEntriesByPropName: this.getEntriesByPropName,
       getTicketPoolById: this.getTicketPoolById,
       ticketDefinitions: this.getTicketDefinitions(poolId),