+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 * Method `getTicketPoolById` from `TicketModalController` is publicly available to it's children.
 * Added `wholeDay` to cart entry.
 
+## 1.4.0
+* Added `getAvailablePoolTicketsQty` to `TicketModalController`.
+* Added `getTicketAvailabilityText` to `TicketModalController`.
+
 ## 1.3.0
 * Added minimal supported node (8.15.0) and npm (6.4.1) versions.